# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Laravel with PostgreSQL - Created Muhamad Ferdinal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/pesanan": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get list of orders",
                "description": "Returns list of orders",
                "operationId": "findAllOrder",
                "responses": {
                    "200": {
                        "description": "Pengambilan Data Pesanan Berhasil!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data Pesanan!"
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create order data",
                "description": "Create order data",
                "operationId": "createOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "pesanan_produk"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID User (Required)",
                                        "type": "uuid"
                                    },
                                    "kode_voucher": {
                                        "description": "Kode Voucher Pesanan (Nullable)",
                                        "type": "string"
                                    },
                                    "tgl_pesanan": {
                                        "description": "Tanggal Pesanan (Nullable)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tgl_pembayaran_lunas": {
                                        "description": "Tenggal Pembayaran Lunas (Nullable)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tgl_dibatalkan": {
                                        "description": "Tanggal Pesanan Dibatalkan (Nullable)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "no_pesanan": {
                                        "description": "Nomor Pesanan (Nullable)",
                                        "type": "string"
                                    },
                                    "pesanan_produk": {
                                        "description": "Array Pesanan Produk (Required)",
                                        "type": "array",
                                        "items": {}
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Pesanan Berhasil Dibuat!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Membuat Data Pesanan!"
                    }
                }
            }
        },
        "/pesanan/{pesanan_id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get single order data",
                "description": "Returns single order data",
                "operationId": "findOneOrder",
                "parameters": [
                    {
                        "name": "pesanan_id",
                        "in": "path",
                        "description": "Pesanan id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengambilan Data Pesanan Berhasil!"
                    },
                    "404": {
                        "description": "Data Pesanan Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Pengambilan Data Pesanan!"
                    }
                }
            },
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update order data",
                "description": "Update order data",
                "operationId": "updateOneOrder",
                "parameters": [
                    {
                        "name": "pesanan_id",
                        "in": "path",
                        "description": "Pesanan id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "ID User (Optional)",
                                        "type": "uuid"
                                    },
                                    "kode_voucher": {
                                        "description": "Kode Voucher Pesanan (Optional)",
                                        "type": "string"
                                    },
                                    "tgl_pesanan": {
                                        "description": "Tanggal Pesanan (Optional)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tgl_pembayaran_lunas": {
                                        "description": "Tenggal Pembayaran Lunas (Optional)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tgl_dibatalkan": {
                                        "description": "Tanggal Pesanan Dibatalkan (Optional)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "no_pesanan": {
                                        "description": "Nomor Pesanan (Optional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data Pesanan Berhasil Diperbarui!"
                    },
                    "404": {
                        "description": "Data Pesanan Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Memperbarui Data Pesanan!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Delete order data",
                "description": "Delete order data",
                "operationId": "deleteOneOrder",
                "parameters": [
                    {
                        "name": "pesanan_id",
                        "in": "path",
                        "description": "Pesanan id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data Pesanan Berhasil Dihapus!"
                    },
                    "404": {
                        "description": "Data Pesanan Tidak Ditemukan!"
                    },
                    "500": {
                        "description": "Terjadi Kesalahan Saat Menghapus Data Pesanan!"
                    }
                }
            }
        },
...
    }
}